fix: #71 Error 1046: No database selected
The MySQL `USE` statement is good only for the connection in which it is
executed. When the connection is recycled or a different connection is
used from the pool managed by sql.DB then the queries don't work.

This patch ensures all queries are executed by referencing the
fully-qualified name of the table: `<schema>.<table>`.

This change breaks usage of the store's `Query()` method depending on
whether the DSN specifies the database name and whether the query string
is using the fully qualified table name.

Signed-off-by: George Aristy <george.aristy@securekey.com>