feat: New command utility to get CSV args
Added functions for retrieving user-set variables from either the command line (via the Cobra library) or from an environment variable that are in CSV format.

The new functions are similar to GetUserSetOptionalVarFromArrayString and GetUserSetVarFromArrayString, except that the command line variable can be in CSV format instead of repeated flags. The behaviour of those existing methods can be somewhat unintuitive since they require their input to be in two different formats depending on whether you're using the command line or an environment variable. These new methods provide more consistency, which makes documentation simpler too.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>